Add files via upload
Spotify Data Analysis
In this collaborative project with my teammates Kalys and Firus, we analyzed music data from the 1960s to the 2010s, focusing on trends that define hit songs. Together, we built a decision tree model that achieved an 80% prediction accuracy for identifying potential hits. This project highlights our collective skills in data cleaning, visualization, and teamwork.